[impl] mst strategy
