+++ lib/bsp/strategy/strategy_executor.dart
@@ -1,4 +1,4 @@
-import 'package:push_puzzle/bsp/mst/mst.dart';
+import 'package:push_puzzle/bsp/strategy/mst.dart';
 import 'package:push_puzzle/bsp/strategy/strategy.dart';
 
 class ContextExecutor {
